requirements.txt
yfinance==0.2.33
kafka-python==2.0.2
pyspark==3.5.0
psycopg2-binary==2.9.9
redis==5.0.1
streamlit==1.29.0
pandas==2.1.4
numpy==1.26.2
python-dotenv==1.0.0
requests==2.31.0
matplotlib==3.8.2
plotly==5.18.0
tabulate==0.9.0
